# ENH: optimize import times