@mrocklin 2017-10-24
I am glad to see progress on this and very glad to see your involvement in `multipledispatch` generally.  However, I am also particularly busy this week and probably won't look at this until next Tuesday at the earliest.  Please don't take this as a sign that I am anything less than very enthusiastic :)