@mrocklin 2017-11-16
In general the approach here seems quite nice.  I can see how it would have a large positive impact without any requirements on the user and without much complexity.

Another option, should you be interested, would be to maintain some data structure so that we could update the ordering cheaply when we find a new signature.  Currently we clear out all intermediate state when we solve.  Presumably there is something we could keep around to make incremental additions efficient.