@mrocklin 2017-12-27
Does this meet your operational performance constraints?