@mrocklin 2017-12-27
You were also concerned about call-time costs.  My hope is that caching handles almost all of these costs, but wanted to check.