@llllllllll 2017-12-27
Ah, thanks for reminding me about the per-call cost. I will run a more serious benchmark using blaze and get back to you.